Move phenotype validation to form for GWAS views.
